requirements.txt
streamlit
pandas
openpyxl
plotly
scikit-learn
numpy
matplotlib

google-generativeai
openai
